+++ CHANGELOG.md
@@ -4,10 +4,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [ita-challenges-frontend-3.4.0-RELEASE] - 2025-09-17
+### [ita-challenges-frontend-3.5.1-RELEASE] - 2025-09-17
 
 ### Added
 - Using the edit service Mentor can edit a current challenge by clicking on edit button (Taiga [#658], PR [#691]).
+### [ita-challenges-frontend-3.5.0-RELEASE] - 2025-09-15
+
+### Added
+- Added update service challenge for mentor, just to modify the old version of the challenge, and save it on the db (Taiga [#659], PR [#690]).
+### [ita-challenges-frontend-3.4.1-RELEASE] - 2025-09-14
+
+### Added
+- Added languages for the error message when submiting new challenge without any selected tags (Taiga [#662], PR [695]).
+
+### [ita-challenges-frontend-3.4.0-RELEASE] - 2025-09-14
+
+### Changed
+- Added provisional error message when submiting new challenge without any selected tags (Taiga [#653], PR [686]).
 
 ### [ita-challenges-frontend-3.3.2-RELEASE] - 2025-07-30